Add hacky auth/auth to all routes

+++ src/database/database_upsun.py
@@ -17,6 +17,7 @@ class APIKey(Base):
     salt = Column(String, nullable=False)
     api_key = Column(String, nullable=False)
     expiration_date = Column(DateTime, nullable=False)
+    role = Column(String, default='user')
 
 # Create an engine
 engine = create_engine(os.getenv('DATABASE_URL'), echo=True, future=True)
@@ -34,12 +35,25 @@ def generate_api_key(length=32):
     api_key = ''.join(secrets.choice(characters) for _ in range(length))
     return api_key
 
-def add_api_key(username, expiration_date):
+def add_api_key(key, expiration_date):
+    # Retrieve username
     session = Session()
+    keys = session.query(APIKey).filter().all()
+    username = "null"
+    role="user"
+    for db_key in keys:
+        found = bcrypt.checkpw(key.encode('utf-8'), db_key.api_key.encode('utf-8'))
+        if found is True:
+            username = db_key.username
+            role = db_key.role
+            break
+    if username in ["null"]:
+        # TODO return a proper error object
+        raise Exception("key has no username.")
     salt = bcrypt.gensalt()
     api_key = generate_api_key()
     hashed = bcrypt.hashpw(api_key.encode('utf-8'), salt)
-    new_key = APIKey(username=username, salt=salt.decode('utf-8'), api_key=hashed.decode('utf-8'), expiration_date=expiration_date)
+    new_key = APIKey(username=username, salt=salt.decode('utf-8'), api_key=hashed.decode('utf-8'), expiration_date=expiration_date, role=role)
     session.add(new_key)
     session.commit()
     session.close()
@@ -47,29 +61,46 @@ def add_api_key(username, expiration_date):
 
 def validate_api_key(api_key):
     session = Session()
-    
-    # First we have to get the salt
     keys = session.query(APIKey).filter().all()
     found = False
     for key in keys:
-        print(f"key: {key}")
-        hashed_api_key = bcrypt.hashpw(api_key.encode('utf-8'), key.salt.encode('utf-8'))
-        key = session.query(APIKey).filter(APIKey.api_key == hashed_api_key.decode('utf-8')).first()
-        if key is not None:
-            found = True
+        found = bcrypt.checkpw(api_key.encode('utf-8'), key.api_key.encode('utf-8'))
+        if found is True:
             break
     session.close()
     return found
 
-def get_api_keys(username):
+def get_api_keys(username, api_key):
     session = Session()
+
     keys = session.query(APIKey).filter(APIKey.username == username).all()
+    found = False
+    for db_key in keys:
+        found = bcrypt.checkpw(api_key.encode('utf-8'), db_key.api_key.encode('utf-8'))
+        if found is True:
+            break
+    if found is False:
+        # TODO return a proper error object
+        raise Exception("API Key belongs to another user.")
+
     session.close()
     return keys
 
-def delete_api_key(username, api_key_id):
+def delete_api_key(key_id, api_key):
     session = Session()
-    key = session.query(APIKey).filter(APIKey.id == api_key_id, APIKey.username == username).first()
+
+    keys = session.query(APIKey).filter().all()
+    username = "null"
+    for db_key in keys:
+        found = bcrypt.checkpw(api_key.encode('utf-8'), db_key.api_key.encode('utf-8'))
+        if found is True:
+            username = db_key.username
+            break
+    if username in ["null"]:
+        # TODO return a proper error object
+        raise Exception("key has no username.")
+
+    key = session.query(APIKey).filter(APIKey.id == key_id and APIKey.username == username).first()
     if key:
         session.delete(key)
         session.commit()
